+++ src/environment/__init__.py
@@ -1,2 +1,9 @@
-"""Environment discovery package."""
+from environment.discovery import EnvironmentDiscovery, save_environment_candidates_json
+from environment.models import EnvironmentCandidate, EnvironmentName
 
+__all__ = [
+    "EnvironmentCandidate",
+    "EnvironmentDiscovery",
+    "EnvironmentName",
+    "save_environment_candidates_json",
+]